Updates
Updates the following:
Moved the Kwh capacity from the sensor.py to the __init__.py so we can declare as global variable

logic for the charge plan that includes cost is now done so it'll pick the cheapest slots for over night charging based upon battery level, this will dynamically update throughout the day every half hour up until midnight.

next steps are to get it to actually run the charging so it'll check the time and select the slot, send the start and end time to the inverter and then change to the next slot.